refactor(control): split tcp flow
